CCarRent/backend/car_rent/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin==2.0.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.1
excel-base==1.0.4
isoweek==1.3.3
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
tzlocal==5.2
whitenoise==6.6.0
xlwt==1.3.0
